refactor(token-controls): lift the catalog filter into the shared library
`filterCatalogOptions` and its render cap move from the Style Library's helpers
into the shared token-control library, and are exported from its barrel.

Both hosts search the same ~1,900-name font catalog: the Style Library's
Typography dropdown today, and the block editor's font-family picker next. A cap
or a match rule that drifted between two copies would make the same query return
different lists on the two screens, which is the kind of difference nobody
reports as a bug because each screen looks reasonable on its own.

Pure move. The module is already React-free and REST-free, so nothing about it
changes beyond its path and its docblock; the only other edits are the import in
`SearchableSelectDropdown` and the one in its test.
